pyproject.toml
[project]
name = "planmode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "glom>=25.12.0",
    "monsterui>=1.0.45",
    "pydantic>=2.12.5",
    "python-fasthtml>=0.12.50",
    "sh>=2.2.2",
    "starlette<0.45.0",
    "typer>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "typer>=0.24.0",
]

[project.scripts]
planmode = "planmode:pm"
pm = "planmode:pm"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.uv]
package = true
